feat: change knobs example to be more realistic
